+++ src/lib/compiler/compileData.ts
@@ -122,6 +122,8 @@ import {
   Avatar,
   ColorCorrectionSetting,
   Font,
+  MonoBGPPalette,
+  MonoOBJPalette,
   Music,
   Palette,
   ProjectResources,
@@ -384,6 +386,9 @@ const precompilePalettes = async (
   const isColor = settings.colorMode !== "mono" || settings.sgbEnabled;
 
   const palettesLookup = indexById(palettes);
+  const defaultBGP = settings.defaultMonoBGP || [0, 1, 2, 3];
+  const defaultOBP0 = settings.defaultMonoOBP0 || [0, 1, 3];
+  const defaultOBP1 = settings.defaultMonoOBP1 || [0, 2, 3];
   const defaultBackgroundPaletteIds =
     settings.defaultBackgroundPaletteIds || [];
   const defaultSpritePaletteIds = settings.defaultSpritePaletteIds || [];
@@ -418,25 +423,52 @@ const precompilePalettes = async (
     };
   };
 
+  const getDMGPalette = <T extends MonoBGPPalette | MonoOBJPalette>(
+    scenePal: T | undefined,
+    defaultPal: T,
+  ): [string, string, string, string] => {
+    const defaultPalette = [
+      "DMG_WHITE",
+      "DMG_LITE_GRAY",
+      "DMG_DARK_GRAY",
+      "DMG_BLACK",
+    ] as [string, string, string, string];
+    if (defaultPal.length === 3 && (!scenePal || scenePal.length === 3)) {
+      // OBJ palette
+      const palette: MonoOBJPalette = scenePal || defaultPal;
+      return [
+        defaultPalette[palette[0]],
+        defaultPalette[palette[0]],
+        defaultPalette[palette[1]],
+        defaultPalette[palette[2]],
+      ];
+    } else if (
+      defaultPal.length === 4 &&
+      (!scenePal || scenePal.length === 4)
+    ) {
+      // BGP palette
+      const palette: MonoBGPPalette = scenePal || defaultPal;
+      return [
+        defaultPalette[palette[0]],
+        defaultPalette[palette[1]],
+        defaultPalette[palette[2]],
+        defaultPalette[palette[3]],
+      ];
+    }
+    return defaultPalette;
+  };
+
   // Background palettes
 
   for (let i = 0; i < scenes.length; i++) {
     const scene = scenes[i];
     const sceneBackgroundPaletteIds = scene.paletteIds || [];
-
     const background = backgrounds[scene.backgroundId];
     if (background?.autoPalettes?.[0]) {
     }
 
     const scenePalette = {
-      dmg: [
-        ["DMG_WHITE", "DMG_LITE_GRAY", "DMG_DARK_GRAY", "DMG_BLACK"] as [
-          string,
-          string,
-          string,
-          string,
-        ],
-      ],
+      dmg: [getDMGPalette(scene.monoBGP, defaultBGP)],
       colors: isColor
         ? [
             getBackgroundPalette(
@@ -512,18 +544,8 @@ const precompilePalettes = async (
 
     const actorsPalette = {
       dmg: [
-        ["DMG_WHITE", "DMG_WHITE", "DMG_LITE_GRAY", "DMG_BLACK"] as [
-          string,
-          string,
-          string,
-          string,
-        ],
-        ["DMG_WHITE", "DMG_WHITE", "DMG_DARK_GRAY", "DMG_BLACK"] as [
-          string,
-          string,
-          string,
-          string,
-        ],
+        getDMGPalette(scene.monoOBP0, defaultOBP0),
+        getDMGPalette(scene.monoOBP1, defaultOBP1),
       ],
       colors: isColor
         ? [